Pass Context instead of Client through click
In order to inject a fake RedashClient and Editor in unit tests, ctx.obj
has to be extended to include both. I borrowed this solution from Stack
Overflow:

https://stackoverflow.com/questions/45938091/multiple-context-objects-in-cli-click

I believe there's a better way to do this - perhaps with pass decorators
- but it's probably a refactor for another day.

This refactor includes a change to Editor's initializer to construct it
with a RedashClient. Additionally, there's a change to cli_runner to
use a MockEditor by default.